I have optimized the different common and non common functions to be indipendent and combined to increase correction effetiency and also make sure they dont have unnecessary repeats of code.
