+++ analyzers/__init__.py
@@ -1,4 +1,4 @@
-from .apt.chroot_analyzer import apt_chroot_analysis
-from .apt.static_analyzer import apt_static_analysis
-from .rpm.rpm_chroot_analyzer import rpm_chroot_analysis
-from .rpm.rpm_static_analyzer import rpm_static_analysis
+from .apt_package_manager.chroot_analyzer import apt_chroot_analysis
+from .apt_package_manager.static_analyzer import apt_static_analysis
+from .rpm_package_manager.rpm_chroot_analyzer import rpm_chroot_analysis
+from .rpm_package_manager.rpm_static_analyzer import rpm_static_analysis
