--- analyzers/pacman/placeholder.py
@@ -1,1 +1,0 @@
-# Working on a working code
